refactor base-1
